tests/unit: move disk_test_utils bodies to .cpp
`disk_test_utils.hpp` was a header-only helper included by 42 unit tests.
Its three inline functions transitively pulled `rocksdb/utilities/transaction_db.h`
and `dbms/constants.hpp` into every consumer, costing parse time in each TU.

Move the bodies into a new `disk_test_utils.cpp` built as a small helper
library; reduce the header to declarations plus a forward-decl of
`rocksdb::TransactionDB`. `storage/v2/disk/storage.hpp` stays in the header
because every consumer uses `DiskStorage` directly.

Link the new `disk_test_utils` library into the 42 consuming unit tests.
